+++ Src/Roflcopter.Plugin/MismatchedFileNames/OptionsPages/MismatchedFileNamesOptionsPage.cs
@@ -14,7 +14,10 @@ namespace Roflcopter.Plugin.MismatchedFileNames.OptionsPages
         typeofIcon: typeof(AlteringFeatuThemedIcons.FileHeaderText),
         ParentId = CodeInspectionPage.PID)]
     [ExcludeFromCodeCoverage /* manually tested UI code */]
+#pragma warning disable 618
+    // TODO after dropping 20181 support: Refactor to BeSimpleOptionsPage
     public class MismatchedFileNamesOptionsPage : SimpleOptionsPage
+#pragma warning restore 618
     {
         private const string OptionsPageId = nameof(MismatchedFileNamesOptionsPage);