+++ Src/Roflcopter.Plugin/TodoItems/Presentation/ZoneMarker.cs
@@ -1,9 +1,12 @@
 ﻿using JetBrains.Application.BuildScript.Application.Zones;
+#if RS20181
+using IUIInteractiveEnvZone = JetBrains.Application.BuildScript.Application.Zones.IUIInteractiveZone;
+#endif
 
 namespace Roflcopter.Plugin.TodoItems.Presentation
 {
     [ZoneMarker]
-    public class ZoneMarker : IRequire<IUIInteractiveZone>
+    public class ZoneMarker : IRequire<IUIInteractiveEnvZone>
     {
     }
 }